+++ packages/graphql-hooks/test/unit/useClientRequest.test.tsx
@@ -1,7 +1,6 @@
 import { act, renderHook } from '@testing-library/react-hooks'
-import T from 'prop-types'
 import React from 'react'
-import { ClientContext, useClientRequest } from '../../src'
+import { ClientContext, useClientRequest, UseClientRequestOptions } from '../../src'
 
 let mockClient
 
@@ -10,9 +9,6 @@ const Wrapper = props => (
     {props.children}
   </ClientContext.Provider>
 )
-Wrapper.propTypes = {
-  children: T.node
-}
 
 const TEST_QUERY = `query Test($limit: Int) {
   tests(limit: $limit) {
@@ -53,7 +49,7 @@ describe('useClientRequest', () => {
   })
 
   it('uses a passed client', async () => {
-    const mockClient2 = {
+    const mockClient2: any = {
       mutationsEmitter: {
         emit: jest.fn()
       },
@@ -66,7 +62,7 @@ describe('useClientRequest', () => {
       },
       request: jest.fn().mockResolvedValue({ data: 'data' })
     }
-    const options = { isMutation: false, client: mockClient2 }
+    const options: UseClientRequestOptions = { isMutation: false, client: mockClient2 };
     let fetchData
     renderHook(() => ([fetchData] = useClientRequest(TEST_QUERY, options)), {
       wrapper: Wrapper
@@ -78,6 +74,14 @@ describe('useClientRequest', () => {
       { query: TEST_QUERY },
       options
     )
+  })
+
+  it('returns an error if there is no client available', () => {
+    const options: UseClientRequestOptions = { isMutation: false }
+
+    const { result } = renderHook(() => useClientRequest(TEST_QUERY, options))
+
+    expect(result.error?.message).toEqual( 'A client must be provided in order to use the useClientRequest hook.')
   })
 
   it('resets data when reset function is called', async () => {
@@ -315,10 +319,10 @@ describe('useClientRequest', () => {
   })
 
   it('throws if the supplied query is not a string', () => {
-    const rendered = renderHook(() => useClientRequest({}), {
+    const rendered = renderHook(() => useClientRequest({} as any), {
       wrapper: Wrapper
     })
-    expect(rendered.result.error.message).toMatch(
+    expect(rendered?.result?.error?.message).toMatch(
       /^Your query must be a string/
     )
   })
@@ -665,7 +669,7 @@ describe('useClientRequest', () => {
             ([fetchData] = useClientRequest(TEST_QUERY, {
               variables: { limit: 10 },
               updateData: 'do I look like a function to you?'
-            })),
+            } as any)),
           { wrapper: Wrapper }
         )
 
@@ -717,7 +721,7 @@ describe('useClientRequest', () => {
 
     describe('memoisation', () => {
       it('returns the same function on every render if query and options remain the same', () => {
-        const fetchDataArr = []
+        const fetchDataArr: any[] = []
         const { rerender } = renderHook(
           () => {
             const [fetchData] = useClientRequest(TEST_QUERY, {
@@ -735,7 +739,7 @@ describe('useClientRequest', () => {
       })
 
       it('returns a new function if query or options change', () => {
-        const fetchDataArr = []
+        const fetchDataArr: any[] = []
 
         const { rerender } = renderHook(
           ({ variables }) => {
@@ -816,7 +820,7 @@ describe('useClientRequest', () => {
     })
 
     it('emits an event when a mutation returns its result', async () => {
-      const mockClient = {
+      const mockClient: any = {
         mutationsEmitter: {
           emit: jest.fn()
         },
@@ -829,7 +833,7 @@ describe('useClientRequest', () => {
         },
         request: jest.fn().mockResolvedValue({ data: 'data' })
       }
-      const options = {
+      const options: UseClientRequestOptions = {
         isMutation: true,
         client: mockClient,
         variables: {